+++ src/Compute/Compute/Strategies/ComputeRp/VirtualMachineScaleSetStrategy.cs
@@ -71,7 +71,10 @@ namespace Microsoft.Azure.Commands.Compute.Strategies.ComputeRp
             string imageReferenceId,
             Dictionary<string, List<string>> auxAuthHeader,
             string diskControllerType,
-            string sharedImageGalleryId
+            string sharedImageGalleryId,
+            string securityType = null,
+            bool? enableVtpm = null,
+            bool? enableSecureBoot = null
             )
             => Strategy.CreateResourceConfig(
                 resourceGroup: resourceGroup,
@@ -96,7 +99,13 @@ namespace Microsoft.Azure.Commands.Compute.Strategies.ComputeRp
                         PlatformFaultDomainCount = platformFaultDomainCount,
                         VirtualMachineProfile = new VirtualMachineScaleSetVMProfile
                         {
-                            SecurityProfile = (encryptionAtHost == true) ? new SecurityProfile(encryptionAtHost: encryptionAtHost) : null,
+                            SecurityProfile = (encryptionAtHost == true || enableVtpm != null || enableSecureBoot != null || securityType != null) 
+                            ? new SecurityProfile
+                            {
+                                EncryptionAtHost = encryptionAtHost,
+                                UefiSettings = (enableVtpm != null || enableSecureBoot != null) ? new UefiSettings(enableSecureBoot, enableVtpm) : null,
+                                SecurityType = securityType,
+                            } : null,
                             OsProfile = new VirtualMachineScaleSetOSProfile
                             {
                                 ComputerNamePrefix = name.Substring(0, Math.Min(name.Length, 9)),
@@ -199,7 +208,10 @@ namespace Microsoft.Azure.Commands.Compute.Strategies.ComputeRp
             int? platformFaultDomainCount,
             string edgeZone,
             string orchestrationMode,
-            string capacityReservationId
+            string capacityReservationId,
+            bool? enableVtpm = null,
+            bool? enableSecureBoot = null,
+            string securityType = null
             )
             => Strategy.CreateResourceConfig(
                 resourceGroup: resourceGroup,
@@ -219,7 +231,13 @@ namespace Microsoft.Azure.Commands.Compute.Strategies.ComputeRp
                     PlatformFaultDomainCount = platformFaultDomainCount,
                     VirtualMachineProfile = new VirtualMachineScaleSetVMProfile
                     {
-                        SecurityProfile = (encryptionAtHost == true) ? new SecurityProfile(encryptionAtHost: encryptionAtHost) : null,
+                        SecurityProfile = (encryptionAtHost == true || enableVtpm != null || enableSecureBoot != null || securityType != null) 
+                        ? new SecurityProfile
+                        {
+                            EncryptionAtHost = encryptionAtHost,
+                            UefiSettings = (enableVtpm != null || enableSecureBoot != null) ? new UefiSettings(enableSecureBoot, enableVtpm) : null,
+                            SecurityType = securityType,
+                        } : null,
                         OsProfile = new VirtualMachineScaleSetOSProfile
                         {
                             ComputerNamePrefix = name.Substring(0, Math.Min(name.Length, 9)),