+++ src/Ssh/Ssh/ChangeLog.md
@@ -19,6 +19,8 @@
 -->
 
 ## Upcoming Release
+* Support updating Service Configuration for Arc resources at runtime.
+    - Owners/Contributors can change what port is allowed for SSH connection at runtime by providing the -Port parameter and confirming the operation when prompted.
 
 ## Version 0.1.2
 * The SSH Proxy required for connection to Arc resources must be installed by the user as part of the Az.Ssh.ArcProxy PowerShell module
@@ -26,7 +28,7 @@
     - The proxy files were previously downloaded by the cmdlet at runtime.
 
 ## Version 0.1.1
-* Added support for the following Resource Types: 
+* Added support for the following Resource Types:
     - Microsoft.ConnectedVMwarevSphere/virtualMachines
     - Microsoft.ScVmm/virtualMachines
     - Microsoft.AzureStackHCI/virtualMachines